# swagger/swagger.yaml
swagger: "2.0"
info:
  description: "You can visit out RiderTrack page at [https://rider-track-dev.herokuapp.com]" 
  version: "1.0.0"
  title: "Ridertrack Swagger"
host: "rider-track-dev.herokuapp.com"
basePath: "/api"
...
schemes:
- "https"
paths:
...
  /events/{eventId}/route:
    get:
      tags:
      - "route"
      summary: "It returns the route of the requested event."
      description: "It returns the route of the requested event."
      produces:
      - "application/json"
      parameters:
      - name: eventId
        in: path
        type: "string"
        required: true
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/Route"
        400:
          description: "Error message"
    post:
      tags:
      - "route"
      summary: "It creates the route for the given event."
      description: "It creates the route for the given event."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: eventId
        type: "string"
        required: true
      - in: body
        name: body
        description: "Object of coordinates that make a route"
        required: true
        schema:
          description: ""
      responses:
        400:
          description: "errors"
        200:
          description: "Route successfully created"
          schema:
            $ref: "#/definitions/Coordinates"
    put:
      tags:
      - "route"
      summary: "Update route of an event"
      description: "It updates the route for the given event."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: eventId
        in: path
        type: string
        required: true
      - in: body
        name: body
        description: "Object of coordinates that make a route"
        required: true
        schema:
          description: ""
      responses:
        200:
          description: "Route successfully updated"
          schema:
            $ref: "#/definitions/Coordinates"
        400:
          description: "Error message"
    delete:
      tags:
      - "route"
      summary: "Delete a route"
      description: "Delete route of an event by given Id"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: "eventId"
        type: "string"
        required: true
      responses:
        200:
          description: "Route successfully deleted"
        400:
          description: "Error message"
...
  /enrollments:
    get:
      tags:
      - "enrollment"
      summary: "It returns the list of all the enrollments."
      description: "It returns the list of all the enrollments. It accepts query params for filtering the enrollments: eventId, userId."
      produces:
      - "application/json"
      parameters:
      - name: "eventId"
        in: "query"
        description: "Filter by eventId"
        type: "string"
      - name: "userId"
        in: "query"
        description: "Filter by userId"
        type: "string"
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/Enrollment'
        400:
          description: "Error message"
    post:
      tags:
      - "enrollment"
      summary: "It creates the enrollment passed in the body after checking the user is logged in."
      description: "It creates the enrollment passed in the body after checking the user is logged in. It returns the detail of the enrollment just created."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Enrollment object with parameters that you want to create"
        required: true
        schema:
          $ref: "#/definitions/Enrollment"
      responses:
        200:
          description: "User enrolled successfully!"
          schema:
            $ref: '#/definitions/Enrollment'
        400:
          description: "Error message"
  /enrollments/{eventId}:
    get:
      tags:
      - "enrollment"
      summary: "It return the enrollments of the requested eventId"
      description: "It return the enrollments of the requested eventId"
      produces:
      - "application/json"
      parameters:
      - name: "eventId"
        in: "path"
        type: "string"
        required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/Enrollment'
        400:
          description: "Error message"
...
  /enrollments/{eventId}/{userId}:
    delete:
      tags:
      - "enrollment"
      summary: " It deletes the given enrollment by eventId and userId"
      description: "It deletes the given enrollment by eventId and userId. Can be called only by the given user if he/she is enrollment on the event. This will delete permanently everything related to it."
      produces:
      - "application/json"
      parameters:
      - name: "eventId"
        in: "path"
        type: "string"
        required: true
      - name: "userId"
        in: "path"
        type: "string"
        required: true
      responses:
        200:
          description: "Enrollment successfully deleted"
          schema:
            $ref: '#/definitions/Enrollment'
        400:
          description: "Error message"
        401:
          description: "You are not authorized to delete this enrollment."
...
  /admin/events/{eventId}/route:
    get:
      tags:
      - admin
      summary: "It returns the route of the requested event."
      description: "It returns the route of the requested event."
      produces:
      - "application/json"
      parameters:
      - name: eventId
        in: path
        type: "string"
        required: true
      responses:
        200:
          description: ""
          schema:
            $ref: "#/definitions/Route"
        400:
          description: "Error message"
    post:
      tags:
      - admin
      summary: "It creates the route for the given event."
      description: "It creates the route for the given event."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: eventId
        type: "string"
        required: true
      - in: body
        name: body
        description: "Object of coordinates that make a route"
        required: true
        schema:
          description: ""
      responses:
        400:
          description: "errors"
        200:
          description: "Route successfully created"
          schema:
            $ref: "#/definitions/Coordinates"
    put:
      tags:
      - admin
      summary: "Update route of an event"
      description: "It updates the route for the given event."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: eventId
        in: path
        type: string
        required: true
      - in: body
        name: body
        description: "Object of coordinates that make a route"
        required: true
        schema:
          description: ""
      responses:
        200:
          description: "Route successfully updated"
          schema:
            $ref: "#/definitions/Coordinates"
        400:
          description: "Error message"
    delete:
      tags:
      - admin
      summary: "Delete a route"
      description: "Delete route of an event by given Id"
      produces:
      - "application/json"
      parameters:
      - in: path
        name: "eventId"
        type: "string"
        required: true
      responses:
        200:
          description: "Route successfully deleted"
        400:
          description: "Error message"
...
  /admin/enrollments:
    get:
      tags:
      - admin
      summary: "It returns the list of all the enrollments."
      description: "It returns the list of all the enrollments. It accepts query params for filtering the enrollments: eventId, userId."
      produces:
      - "application/json"
      parameters:
      - name: "eventId"
        in: "query"
        description: "Filter by eventId"
        type: "string"
      - name: "userId"
        in: "query"
        description: "Filter by userId"
        type: "string"
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/Enrollment'
        400:
          description: "Error message"
    post:
      tags:
      - admin
      summary: "It creates the enrollment passed in the body after checking the user is logged in."
      description: "It creates the enrollment passed in the body after checking the user is logged in. It returns the detail of the enrollment just created."
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "Enrollment object with parameters that you want to create"
        required: true
        schema:
          $ref: "#/definitions/Enrollment"
      responses:
        200:
          description: "User enrolled successfully!"
          schema:
            $ref: '#/definitions/Enrollment'
        400:
          description: "Error message"
  /admin/enrollments/{eventId}:
    get:
      tags:
      - admin
      summary: "It return the enrollments of the requested eventId"
      description: "It return the enrollments of the requested eventId"
      produces:
      - "application/json"
      parameters:
      - name: "eventId"
        in: "path"
        type: "string"
        required: true
      responses:
        200:
          description: ""
          schema:
            $ref: '#/definitions/Enrollment'
        400:
          description: "Error message"
  /admin/enrollments/{enrollmentId}:
    put:
      tags:
      - admin
      summary: " It updates the fields passed in the body of the given enrollmentId"
      description: " It updates the fields passed in the body of the given enrollmentId"
      produces:
      - "application/json"
      parameters:
      - name: "enrollmentId"
        in: "path"
        type: "string"
        required: true
      - in: "body"
        name: "body"
        required: true
        schema:
          description: ""
      responses:
        200:
          description: "Enrollment successfully updated"
          schema:
            $ref: '#/definitions/Enrollment'
        400:
          description: "Error message"
        401:
          description: "You are not authorized to update this enrollment."
  /admin/enrollments/{eventId}/{userId}:
    delete:
      tags:
      - admin
      summary: " It deletes the given enrollment by eventId and userId"
      description: "It deletes the given enrollment by eventId and userId. Can be called only by the given user if he/she is enrollment on the event. This will delete permanently everything related to it."
      produces:
      - "application/json"
      parameters:
      - name: "eventId"
        in: "path"
        type: "string"
        required: true
      - name: "userId"
        in: "path"
        type: "string"
        required: true
      responses:
        200:
          description: "Enrollment successfully deleted"
          schema:
            $ref: '#/definitions/Enrollment'
        400:
          description: "Error message"
        401:
            description: "You are not authorized to delete this enrollment."
definitions:
...
  Route:
    type: "object"
    properties:
      _id:
        type: "string"
      eventId:
        type: "string"
      coordinates:
        $ref: "#/definitions/Coordinates"
  Coordinates:
    type: "object"
    properties:
      lat:
        type: "number"
      lon:
        type: "number"
  Enrollment:
    type: "object"
    properties:
      _id:
        type: "string"
      eventId:
        type: "string"
      userId:
        type: "string"